backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlmodel==0.0.24
pytest==8.4.1
httpx==0.28.1
openai==1.65.1
python-dotenv==1.0.1
PyJWT==2.10.1
